policy: refactor `main.rs` into `lib.rs`
Currently, the `linkerd-policy-controller` crate contains a `lib.rs` and
a `main.rs`. The `lib.rs` defines some of the more generic interfaces
into the policy-controller's application logic, but a substantial amount of
process startup (e.g. handling CLI arguments, configuring clients and
servers, spawning tasks, etc) is performed in the `main` function in
`main.rs`.

This factoring feels somewhat awkward. If it's not really possible to
run a policy-controller process without code that only exists in
`main.rs`, why provide a `lib.rs` interface at all? We may as well move
all of the code necessary to run a policy-controller into the library
crate, and have a very simple `main.rs` that simply calls into the
library's top-level run function. In the future, this could potentially
enable end-to-end testing by running multiple policy-controller binaries
in a single process, or allow downstream code to run the
policy-controller as part of other software that interfaces with
Linkerd.

This branch refactors the policy-controller to move all the application
logic currently defined in `main.rs` into the `lib.rs` as a public `run`
method on the CLI arguments type, and a free function `run()` that
parses CLI args and then runs the controller.

The `main.rs` file now need only contain a three-line `main` function
that just does this:

```rust
fn main() -> anyhow::Result<()> {
    linkerd_policy_controller::run()
}
```

(along with some crate-level attributes and the `static` for using
`jemalloc` as the system allocator)
